worker/requirements.txt
firebase_functions~=0.1.0
firebase_admin~=6.2.0
google-cloud-firestore~=2.11.1
joblib~=1.3.2
scikit-learn~=1.3.0
# Note: Install PyTorch with ROCm support manually:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm6.2
torch>=2.5.0
torchvision>=0.20.0
Pillow~=10.0.0
transformers>=4.50.0
tokenizers>=0.22.0
huggingface_hub>=0.20.0
accelerate>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
timm>=1.0.0  # Required for PixAI Tagger (EVA02 encoder)